feat: добавлена возможность редактировать профиль, добавлять, лайкать, удалять и подтверждать удаление карточки, проведён рефакторинг кода

+++ src/utils/Api.js
@@ -90,24 +90,24 @@ class Api extends React.Component {
     .then(this._checkResponse);
   }
 
-  likeCard(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
-      method: 'PUT',
-      headers: {
-        authorization: this._authorization
-      }
-    })
-    .then(this._checkResponse);
-  }
-
-  removeLike(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
-      method: 'DELETE',
-      headers: {
-        authorization: this._authorization
-      }
-    })
-    .then(this._checkResponse);
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+        method: 'PUT',
+        headers: {
+          authorization: this._authorization
+        }
+      })
+      .then(this._checkResponse);
+    } else {
+      return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+        method: 'DELETE',
+        headers: {
+          authorization: this._authorization
+        }
+      })
+      .then(this._checkResponse);
+    }
   }
 }
